backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
mysql-connector-python==8.3.0
feedparser==6.0.11
openai==1.12.0
anthropic==0.18.0
python-dotenv==1.0.1
httpx==0.27.0
apscheduler==3.10.4
